UI Changes
In the UnitStats script: Changed all the stats variables to be public. Made a constructor that takes all the variables that will get displayed.

In the Player Controller script: Added a function UpdateUnitDisplay that uses TMP_Text to display information from a unit to text on the screen. Modified the Input Controller so that the when a left click happens a dummy unit's information appears.

In the Game Manager script: Added on to the GetUnitOnTile to return a dummy unit's stats. Also changed the parameters type to be of Vector3 instead of Vector3Int (Can be changed back, used to just test clicking anywhere).

General Script: Changed the General : UnitStats to General : MonoBehaviour because it invoked an error while testing. This can be changed back.